feat: refactor `toc` component & update props
