+++ src/lib/components/toc/index.ts
@@ -1,0 +1,1 @@
+export { default as TOC } from "./TOC.svelte";
